// packages/shared/src/types.ts
export type Direction = "ltr" | "rtl";
export type Orientation = "horizontal" | "vertical";
export type Side = "top" | "right" | "bottom" | "left";
export type Align = "start" | "center" | "end";
export type ProgressState = "indeterminate" | "complete" | "loading";
export type HighlightingDirection =
  | "next"
  | "prev"
  | "first"
  | "last"
  | "selected";